live-vm-common.bbclass: conditionally to add kernel and initrd to efi.img of ISO live
Add var-EFI_ISO_KERNEL, if ${EFI_ISO_KERNEL} != "1", do not add kernel and
initrd to efi.img of ISO live.

For grub-efi's ISO live, the kernel and initrd is not necessary in efi.img,
because grub-efi's bootx64.efi sets `search.file ($cmdpath)/EFI/BOOT/grub.cfg root'
to locate kernel and initrd in ISO live image.

It helps while the size of initrd is big to break the limit of efi.img in which
the filesystem type is FAT

Also clean up ${EFIIMGDIR} before populating

Set EFI_ISO_KERNEL ??= "1" by default which adds kernel and initrd to efi.img
as usual

Issue: LIN1024-7994
(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 20250328

+++ meta/classes-recipe/live-vm-common.bbclass
@@ -54,13 +54,21 @@ efi_populate_common() {
         fi
 }
 
+# Kernel(with initrd) image in efi.img of ISO live
+EFI_ISO_KERNEL ??= "1"
+
 efi_iso_populate() {
         iso_dir=$1
         efi_populate $iso_dir
+
+        rm -rf ${EFIIMGDIR}
         # Build a EFI directory to create efi.img
         mkdir -p ${EFIIMGDIR}/${EFIDIR}
         cp $iso_dir/${EFIDIR}/* ${EFIIMGDIR}${EFIDIR}
-        cp $iso_dir/${KERNEL_IMAGETYPE} ${EFIIMGDIR}
+        if [ "${EFI_ISO_KERNEL}" = "1" ]; then
+                cp $iso_dir/${KERNEL_IMAGETYPE} ${EFIIMGDIR}
+        fi
+
 
         EFIPATH=$(echo "${EFIDIR}" | sed 's/\//\\/g')
 
@@ -70,7 +78,7 @@ efi_iso_populate() {
             printf 'fs0:%s\%s\n' "$EFIPATH" "${EFI_BOOT_IMAGE}" >${EFIIMGDIR}/startup.nsh
         fi
 
-        if [ -f "$iso_dir/initrd" ] ; then
+        if [ -f "$iso_dir/initrd" -a "${EFI_ISO_KERNEL}" = "1" ] ; then
                 cp $iso_dir/initrd ${EFIIMGDIR}
         fi
 }
